Fixed a bunch of bugs related to removal/readding pipes
There was honestly a bunch of issues with deleting WorldObject in general. When you removed a world object, it got removed from the world and the project, but lots of things still held on to references to the object. This was never really an issue before because the object wasn't getting queue_freed() is was just getting removed from the world.